@@ -11,7 +11,9 @@ RCT_EXPORT_MODULE();
 
 RCT_EXPORT_METHOD(immediatePhoneCall:(NSString *)number)
 {
-    [[UIApplication sharedApplication] openURL:[NSURL URLWithString:[NSString stringWithFormat:@"tel:%@", number]]];
+    dispatch_async(dispatch_get_main_queue(), ^{
+        [[UIApplication sharedApplication] openURL:[NSURL URLWithString:[NSString stringWithFormat:@"tel:%@", number]]];
+    });
 };
 
 @end
